alter quota from int to float

+++ internal/database/users.go
@@ -4,7 +4,7 @@ type User struct {
 	Id                  int     `json:"id"`
 	Identity            string  `json:"identity" validate:"required"`
 	Name                string  `json:"name" validate:"required,min=1,max=64"`
-	Quota               int     `json:"quota" validate:"min=0"`
+	Quota               float64 `json:"quota" validate:"min=0"`
 	Used                float64 `json:"used" validate:"min=0"`
 	UsedBytes           int64   `json:"used_bytes" validate:"min=0"`
 	Enabled             bool    `json:"enabled"`
